Task 39 (direction Y) followup: retract leftover int64 user-id locals in Mfa/Session controllers to int32
Phase 1.5d assumed indirect consumers (SessionController et al.) were unaffected by the int64->int32 interface retraction; they compiled fine but kept int64 locals feeding the now-int32 interfaces, producing 5 C4244 narrowing warnings.

- MfaController: userId/fallbackUserId/clearUserId int64_t->int32_t, std::stoll->std::stoi (out_of_range now rejected as invalid MFA session by the existing catch)

- SessionController: onValidated lambda param int64_t->int32_t; drop the internalIdInt32 narrowing cast and the static_cast<int64_t> widening in the legacy branch; refresh the three stale 'widened bridge' comments

Verified: full Release rebuild green, C4244 count 0 (was 5), LNK4006 still 0.

+++ libs/drogon/src/controllers/MfaController.cc
@@ -689,10 +689,13 @@ void MfaController::verifyLogin(
     // needed here.
     if (mfaService_ && userRepo_)
     {
-        int64_t userId = 0;
+        // int32: user ids are int32 end-to-end (Task 39 direction Y, DB int4);
+        // std::stoi throws out_of_range for values beyond int32, which the
+        // catch below already treats as an invalid MFA session.
+        int32_t userId = 0;
         try
         {
-            userId = std::stoll(mfaToken);
+            userId = std::stoi(mfaToken);
         }
         catch (const std::exception &)
         {
@@ -753,10 +756,10 @@ void MfaController::verifyLogin(
         return;
     }
 
-    int64_t fallbackUserId = 0;
+    int32_t fallbackUserId = 0;
     try
     {
-        fallbackUserId = std::stoll(mfaToken);
+        fallbackUserId = std::stoi(mfaToken);
     }
     catch (const std::exception &)
     {
@@ -922,10 +925,10 @@ void MfaController::verifyLogin(
                                       };
                                       if (clearDb)
                                       {
-                                          int64_t clearUserId = 0;
+                                          int32_t clearUserId = 0;
                                           try
                                           {
-                                              clearUserId = std::stoll(mfaToken);
+                                              clearUserId = std::stoi(mfaToken);
                                           }
                                           catch (const std::exception &)
                                           {
